=== ChangeSet ===
Refactored Unlocking of the device so it can unlock most of the devices.
Added functionality for disabling and reenabling the keyguard of the device.
Added functionality for getting running tasks id on the device.
Added functionality for bringing given task to the foreground of the device.
Added functionality to wait for the given task to be updated to given possition.

=== Initial Imports ===

Integration tests for testing the new functionality.
Added enums in Routing action and Service Request.
Added manifest permission in order to use the new functionality.

[https://inttools.musala.com/reviewboard/r/121/ $121] see #3659.

git-svn-id: https://trac.musala.com/ARDC/repo@10433 34708651-a5da-c34d-b20c-10ddab38317d